Improve button styling and form validation
New-transaction deposit/withdraw buttons are now outline-buttons, to be congruent
with styling in edit-transaction form.

Color changes:
Deposit: 'success' -> 'secondary'
Edit: 'secondary' -> 'warning'

Deletion and withdrawal are still both red, which is not ideal. It's still an
improvement, however, since withdraw is now only an outline button. What's more
problematic is that the withdraw button and the form validation tooltips have the
same color. I'll call this a partial success for #34, but the issue is not resolved.

Resolve #9